refactor: rename integrations package as ee
