+++ apps/erp/app/routes/api+/integrations.onshape.d.$did.versions.ts
@@ -1,6 +1,6 @@
 import { requirePermissions } from "@carbon/auth/auth.server";
-import { Onshape as OnshapeConfig } from "@carbon/integrations";
-import { OnshapeClient } from "@carbon/integrations/onshape";
+import { Onshape as OnshapeConfig } from "@carbon/ee";
+import { OnshapeClient } from "@carbon/ee/onshape/lib";
 import type { ShouldRevalidateFunction } from "@remix-run/react";
 import type { LoaderFunctionArgs } from "@vercel/remix";
 import { json } from "@vercel/remix";
