+++ apps/erp/app/routes/api+/integrations.onshape.sync.ts
@@ -1,6 +1,6 @@
 import { getCarbonServiceRole } from "@carbon/auth";
 import { requirePermissions } from "@carbon/auth/auth.server";
-import { onShapeDataValidator } from "@carbon/integrations/onshape";
+import { onShapeDataValidator } from "@carbon/ee/onshape/lib";
 import { FunctionRegion } from "@supabase/supabase-js";
 import { json, type ActionFunctionArgs } from "@vercel/remix";
