+++ apps/erp/app/routes/api+/integrations.slack.install.ts
@@ -1,5 +1,5 @@
 import { requirePermissions } from "@carbon/auth/auth.server";
-import { getSlackInstallUrl } from "@carbon/integrations/slack.server";
+import { getSlackInstallUrl } from "@carbon/ee/slack.server";
 import { json, type LoaderFunctionArgs } from "@vercel/remix";
 
 export const config = {
