+++ apps/erp/app/routes/api+/integrations.slack.interactive.ts
@@ -5,7 +5,7 @@ import {
   createSlackWebClient,
   getCarbonEmployeeFromSlackId,
   getSlackIntegrationByTeamId,
-} from "@carbon/integrations/slack.server";
+} from "@carbon/ee/slack.server";
 import type { SupabaseClient } from "@supabase/supabase-js";
 import { FunctionRegion } from "@supabase/supabase-js";
 import { json, type ActionFunctionArgs } from "@vercel/remix";
