+++ apps/erp/app/routes/x+/settings+/integrations.deactivate.$id.tsx
@@ -1,7 +1,7 @@
 import { assertIsPost, error, success } from "@carbon/auth";
 import { requirePermissions } from "@carbon/auth/auth.server";
 import { flash } from "@carbon/auth/session.server";
-import { integrations as availableIntegrations } from "@carbon/integrations";
+import { integrations as availableIntegrations } from "@carbon/ee";
 import { redirect } from "@remix-run/react";
 import type { ActionFunctionArgs } from "@vercel/remix";
 import { deactivateIntegration } from "~/modules/settings/settings.server";
