+++ apps/erp/app/routes/x+/settings+/integrations.tsx
@@ -1,7 +1,7 @@
 import { error } from "@carbon/auth";
 import { requirePermissions } from "@carbon/auth/auth.server";
 import { flash } from "@carbon/auth/session.server";
-import { integrations as availableIntegrations } from "@carbon/integrations";
+import { integrations as availableIntegrations } from "@carbon/ee";
 import { Outlet, useLoaderData } from "@remix-run/react";
 import { redirect, type LoaderFunctionArgs } from "@vercel/remix";
 import { IntegrationsList, getIntegrations } from "~/modules/settings";
