+++ packages/ee/.eslintrc.cjs
