+++ packages/ee/package.json
@@ -1,10 +1,10 @@
 {
-  "name": "@carbon/integrations",
+  "name": "@carbon/ee",
   "version": "*",
   "type": "module",
   "exports": {
     ".": "./src/index.ts",
-    "./exchange-rates/exchange-rates.server": "./src/exchange-rates/exchange-rates.server.ts",
+    "./exchange-rates.server": "./src/exchange-rates/exchange-rates.server.ts",
     "./notifications": "./src/notifications/index.ts",
     "./onshape": "./src/onshape/lib/index.ts",
     "./paperless-parts": "./src/paperless-parts/lib/index.ts",
@@ -22,6 +22,8 @@
     "test": "jest"
   },
   "dependencies": {
+    "@carbon/auth": "*",
+    "@carbon/react": "*",
     "@carbon/utils": "*",
     "@slack/bolt": "~4.4.0",
     "@slack/oauth": "~3.0.3",
