+++ packages/ee/src/exchange-rates/config.tsx
