+++ packages/ee/src/notifications/registry.ts
