+++ packages/ee/src/paperless-parts/config.tsx
