+++ packages/ee/src/slack/config.tsx
