+++ packages/ee/src/slack/lib/client.ts
