+++ packages/ee/src/slack/lib/index.ts
