+++ packages/ee/src/slack/lib/service.ts
