+++ packages/ee/src/types.ts
