+++ packages/jobs/trigger/paperless-parts.ts
@@ -1,5 +1,8 @@
 import { getCarbonServiceRole } from "@carbon/auth";
 import type { Database } from "@carbon/database";
+import type { SupabaseClient } from "@supabase/supabase-js";
+import { task } from "@trigger.dev/sdk/v3";
+import { z } from "zod";
 import {
   getCarbonOrderStatus,
   getCustomerIdAndContactId,
@@ -9,10 +12,7 @@ import {
   getPaperlessParts,
   insertOrderLines,
   OrderSchema,
-} from "@carbon/integrations/paperless-parts";
-import type { SupabaseClient } from "@supabase/supabase-js";
-import { task } from "@trigger.dev/sdk/v3";
-import { z } from "zod";
+} from "../../ee/src/paperless-parts/lib";
 
 const payloadSchema = z.discriminatedUnion("type", [
   z.object({
