+++ packages/jobs/trigger/slack-document-sync.ts
@@ -1,5 +1,8 @@
 import { getCarbonServiceRole, VERCEL_URL } from "@carbon/auth";
 import type { Database } from "@carbon/database";
+import { WebClient } from "@slack/web-api";
+import type { SupabaseClient } from "@supabase/supabase-js";
+import { task } from "@trigger.dev/sdk/v3";
 import {
   type DocumentType,
   type IssueAssignmentUpdate,
@@ -10,10 +13,7 @@ import {
   formatDocumentCreated,
   formatStatusUpdate,
   formatTaskUpdate,
-} from "@carbon/integrations/slack/messages";
-import { WebClient } from "@slack/web-api";
-import type { SupabaseClient } from "@supabase/supabase-js";
-import { task } from "@trigger.dev/sdk/v3";
+} from "../../ee/src/slack/lib/messages";
 
 export const slackDocumentCreated = task({
   id: "slack-document-created",
